(PC-33723)[API] chore: move AdageBaseResponseModel to core/educational/schemas.py
