--- api/src/pcapi/routes/adage/v1/serialization/config.py
@@ -1,9 +1,0 @@
-from datetime import datetime
-
-from pcapi.routes.serialization import BaseModel
-from pcapi.utils.date import format_into_utc_date
-
-
-class AdageBaseResponseModel(BaseModel):
-    class Config:
-        json_encoders = {datetime: format_into_utc_date}
